israel/requirements.txt
feedparser==6.0.11
openai==1.35.3
requests==2.32.3
python-dotenv==1.0.1
